# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API proyecto EVA",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/item_planilla": {
            "get": {
                "tags": [
                    "Item planillas"
                ],
                "summary": "Obtiene una lista de items planilla",
                "operationId": "2363fb7dd14e1f3dddf44b54d474a62a",
                "responses": {
                    "200": {
                        "description": "Lista de de items planilla"
                    }
                }
            },
            "post": {
                "tags": [
                    "Item planillas"
                ],
                "summary": "Crear un nuevo Item planilla",
                "operationId": "7c196d194942d8e71cafb81e959d0f86",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "id_planilla"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Diseño de base de datos"
                                    },
                                    "id_planilla": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item planilla creado con éxito"
                    },
                    "422": {
                        "description": "Problemas con los datos ingresados"
                    }
                }
            }
        },
        "/api/item_planilla/{id}": {
            "get": {
                "tags": [
                    "Item planillas"
                ],
                "summary": "Mostar un item planilla",
                "operationId": "5763b6ac1924c7733c8a936234c6fcb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del item planilla",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del item planilla"
                    },
                    "404": {
                        "description": "item planilla no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Item planillas"
                ],
                "summary": "Actualizar un item planilla",
                "operationId": "f6dea6abd004b3802cfd33393faf1686",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del item planilla",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Diseño de base de datos"
                                    },
                                    "id_planilla": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "item planilla actualizado con éxito"
                    },
                    "404": {
                        "description": "item planilla no encontrado"
                    },
                    "422": {
                        "description": "Problemas con los datos ingresados"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Item planillas"
                ],
                "summary": "Eliminar un item planilla",
                "operationId": "12f94a12466f5ef8b56f018225470f04",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del item planilla",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "item planilla eliminado"
                    },
                    "404": {
                        "description": "item planilla no encontrado"
                    }
                }
            }
        },
        "/api/item_planilla/planilla_seguimiento/{id}": {
            "get": {
                "tags": [
                    "planillas Seguimientos"
                ],
                "summary": "Mostar una lista item planilla por planilla seguimiento",
                "operationId": "d6b401fbc46517167c8466fda645f94c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del empresa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del item planilla por planilla seguimiento"
                    },
                    "404": {
                        "description": "planilla no encontrada"
                    }
                }
            }
        },
...
    },
...
}